requirements.txt
astroid==2.2.5
autopep8==1.4.4
awsebcli==3.15.2
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
DateTime==4.3
et-xmlfile==1.0.1
idna==2.7
isort==4.3.20
jdcal==1.4.1
lazy-object-proxy==1.4.1
lxml==4.3.4
mccabe==0.6.1
openpyxl==2.6.2
pathlib==1.0.1
pdf2image==1.6.0
Pillow==6.1.0
pycodestyle==2.5.0
pylint==2.3.1
pypandoc==1.4
pytesseract==0.2.7
python-dotenv==0.10.3
pytz==2019.1
pywin32==224
requests==2.20.1
semantic-version==2.5.0
six==1.12.0
soupsieve==1.9.2
typed-ast==1.4.0
urllib3==1.25.3
wincertstore==0.2
wrapt==1.11.1
zope.interface==4.6.0
